- fix input signature for unsigned integer - add check-onnx-backend-signature test - reorganize testing code
