+++ src/Runtime/ExecutionSession.cpp.bak
@@ -22,8 +22,6 @@
 #include "llvm/Support/ManagedStatic.h"
 
 namespace onnx_mlir {
-const std::string ExecutionSession::_inputSignatureName = "omInputSignature";
-const std::string ExecutionSession::_outputSignatureName = "omOutputSignature";
 
 ExecutionSession::ExecutionSession(std::string sharedLibPath,
     std::string entryPointName, std::string inputSignatureName) {
@@ -44,21 +42,11 @@ ExecutionSession::ExecutionSession(std::string sharedLibPath,
     throw std::runtime_error(errStr.str());
   }
 
-  _inputSignatureFunc = reinterpret_cast<signatureFuncType>(
-      _sharedLibraryHandle.getAddressOfSymbol(_inputSignatureName.c_str()));
+  _inputSignatureFunc = reinterpret_cast<inputSignatureType>(
+      _sharedLibraryHandle.getAddressOfSymbol(inputSignatureName.c_str()));
   if (!_inputSignatureFunc) {
     std::stringstream errStr;
-    errStr << "Cannot load symbol: '" << _inputSignatureName << "'"
-           << std::endl;
-    throw std::runtime_error(errStr.str());
-  }
-
-  _outputSignatureFunc = reinterpret_cast<signatureFuncType>(
-      _sharedLibraryHandle.getAddressOfSymbol(_outputSignatureName.c_str()));
-  if (!_outputSignatureFunc) {
-    std::stringstream errStr;
-    errStr << "Cannot load symbol: '" << _outputSignatureName << "'"
-           << std::endl;
+    errStr << "Cannot load symbol: 'omInputSignature'" << std::endl;
     throw std::runtime_error(errStr.str());
   }
 }
@@ -81,12 +69,6 @@ ExecutionSession::run(
         omTensorListGetOmtByIndex(wrappedOutput, i), omTensorDestroy));
   }
   return outs;
-}
-
-std::string ExecutionSession::inputSignature() { return _inputSignatureFunc(); }
-
-std::string ExecutionSession::outputSignature() {
-  return _outputSignatureFunc();
 }
 
 ExecutionSession::~ExecutionSession() {
