+++ test/backend/common.py
@@ -1,0 +1,115 @@
+#!/usr/local/bin/python3
+
+##################### common.py ################################################
+#
+# Copyright 2021 The IBM Research Authors.
+#
+################################################################################
+# commom function `compile_model` called by both
+# SignatureExecutionSession and EndiannessAwareExecutionSession
+################################################################################
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
+import sys
+import os
+import onnx
+import subprocess
+from variables import *
+
+# determine the dynamic input and dim
+def determine_dynamic_parameters(test_name):
+    if not args.dynamic:
+        return None
+    # set default value: all inputs, first dimension.
+    # Use this script's arguments '--input' and '--dim' or environment variables
+    # TEST_INPUT and TEST_DIM to control the values.
+    selected_list = {args.input: {args.dim}}
+    test_name_cpu = test_name + "_cpu"
+    if test_name_cpu in test_for_dynamic:
+        if len(test_to_enable_dict[test_name_cpu]) > 1:
+            selected_list = test_to_enable_dict[test_name_cpu].get(DYNAMIC_SHAPE)
+    return selected_list
+
+
+def execute_commands(cmds, dynamic_inputs_dims):
+    if args.verbose:
+        print(" ".join(cmds), file=sys.stderr)
+        print("IMPORTER FORCE DYNAMIC ", dynamic_inputs_dims, file=sys.stderr)
+    my_env = os.environ.copy()
+    env_string = ""
+    if dynamic_inputs_dims is not None:
+        first_input = True
+        for (input_index, dim_indices) in dynamic_inputs_dims.items():
+            if first_input:
+                env_string += str(input_index)
+                first_input = False
+            else:
+                env_string += "|" + str(input_index)
+            first_dim = True
+            for dim_index in dim_indices:
+                if first_dim:
+                    env_string += ":" + str(dim_index)
+                    first_dim = False
+                else:
+                    env_string += "," + str(dim_index)
+        my_env["IMPORTER_FORCE_DYNAMIC"] = env_string
+    subprocess.run(cmds, env=my_env)
+
+
+def compile_model(model, emit):
+    suffix = {"lib": ".so", "jni": ".jar"}
+    target = {"lib": "--EmitLib", "jni": "--EmitJNI"}
+    name = model.graph.name
+
+    # Each model will have its own model_dir. This is necessary for JNI tests
+    # since all the models will extract libmodel.so. So if all the models are
+    # in the same directory, their libmodel.so will trash each other.
+    model_dir = os.path.join(result_dir, name)
+    os.makedirs(model_dir, exist_ok=True)
+
+    print("ONNX_HOME=" + os.getenv("ONNX_HOME"))
+
+    # For real models, the onnx files are downloaded, no need to save again.
+    if (name + "_cpu") in list(map(lambda x: x[0], real_model_tests)):
+        model_name = os.path.join(os.getenv("ONNX_HOME"), "models", name, "model.onnx")
+    # For node models, write the models in memory out to onnx files.
+    else:
+        model_name = os.path.join(model_dir, name + ".onnx")
+        # Save model to disk as model_name.onnx.
+        onnx.save(model, model_name)
+
+    print(
+        (
+            "Success downloading/saving "
+            if os.path.exists(model_name)
+            else "Failure downloading/saving "
+        )
+        + model_name,
+        file=sys.stderr,
+    )
+
+    exec_base = os.path.join(model_dir, name)
+    exec_name = exec_base + suffix[emit]
+
+    # Command
+    command_list = [TEST_DRIVER]
+    if args.mcpu:
+        command_list.append("--mcpu=" + args.mcpu)
+    if args.mtriple:
+        command_list.append("--mtriple=" + args.mtriple)
+    if args.converter or name in test_need_converter:
+        command_list.append("--invokeOnnxVersionConverter=true")
+    command_list.append(target[emit])
+    command_list.append(model_name)
+    command_list.append('-O3')
+    command_list.append("-o=" + exec_base)
+    # Call frontend to process model_name.onnx, bit code will be generated.
+    dynamic_inputs_dims = determine_dynamic_parameters(name)
+    print("cwd: " + os.getcwd(), file=sys.stderr)
+    execute_commands(command_list, dynamic_inputs_dims)
+    if not os.path.exists(exec_name):
+        print("Failed " + TEST_DRIVER + ": " + name, file=sys.stderr)
+    return exec_name
